CmuxGit stage 1: extract local-git metadata service from TabManager
Lift the pure on-disk git-metadata domain out of TabManager into a new
Layer-2 package, CmuxGit, behind a stateless GitMetadataService actor:

- ResolvedGitRepository / GitWorkspaceMetadata value types.
- workspaceMetadata(for:) (branch, dirty, signatures), watchedPaths(for:)
  (incl. submodule gitlinks), repositorySlugs(forDirectory:).
- ~30 nonisolated static parsing helpers: repository resolution (.git file +
  commondir), HEAD/ref reads, git config include/includeIf + glob, index v2/v3/v4
  parse (prefix compression, assume-unchanged/skip-worktree), stat-based dirty
  detection, GitHub slug parsing.

TabManager now injects a GitMetadataService and calls it at the three entry
points (watched paths, repo slugs, initial metadata snapshot); the
GitMetadataService replaces the manual Task.detached offloading. ~1,120 lines
of helpers + value types removed from TabManager (11,506 -> 10,385).

37 package tests (repo resolution, root-escape stop, branch/detached HEAD,
dirty detection, index v2/v4, content-signature stability, watched paths +
submodule gitlinks, slug parsing, full include/includeIf rules) migrated from
the app test target where they tested the now-lifted statics. Remaining
WorkspacePullRequestSidebarTests integration tests exercise the cutover
end-to-end through TabManager. Fixes the carryover stale-comment doc nit
(CmuxWorkspaceGit/WorkspaceGitMetadataWatcherTests -> CmuxGit/CmuxFileWatch).

cmux + cmux-unit build green; swift build + swift test green.
