+++ Packages/CmuxGit/Sources/CmuxGit/GitMetadataService.swift
@@ -1,0 +1,83 @@
+import Foundation
+
+/// Reads a directory's git metadata directly from the on-disk repository,
+/// without spawning a `git` process.
+///
+/// This service does the filesystem work that powers the workspace sidebar's
+/// branch label, dirty indicator, and pull-request badge: resolving the
+/// enclosing repository, parsing `HEAD`/`index`/`config`, and deriving the set
+/// of paths a filesystem watcher should observe to know when that metadata
+/// becomes stale.
+///
+/// It is a stateless `actor` so that this work runs off the caller's actor (the
+/// reads are pure functions of the paths passed in), and so callers `await` a
+/// single seam instead of hand-rolling `Task.detached` offloading. Every read
+/// is a pure function of the directory argument; the actor holds no cached
+/// state.
+///
+/// ```swift
+/// let git = GitMetadataService()
+/// let meta = await git.workspaceMetadata(for: "/path/to/checkout")
+/// if meta.isRepository, meta.isDirty { showDirtyIndicator() }
+/// ```
+public actor GitMetadataService {
+    /// Creates a git-metadata service.
+    public init() {}
+
+    /// Reads a point-in-time git snapshot for `directory`.
+    ///
+    /// Walks upward to the nearest repository, then parses `HEAD`, the `index`,
+    /// and submodule pointers. Returns ``GitWorkspaceMetadata/notARepository``
+    /// when `directory` is not inside a git repository.
+    ///
+    /// - Parameter directory: An absolute path to inspect.
+    /// - Returns: The git metadata for the enclosing repository, or
+    ///   ``GitWorkspaceMetadata/notARepository`` when there is none.
+    public func workspaceMetadata(for directory: String) -> GitWorkspaceMetadata {
+        guard let repository = Self.resolveGitRepository(containing: directory) else {
+            return .notARepository
+        }
+        let trackedChanges = Self.gitTrackedChangesSnapshot(repository: repository)
+        return GitWorkspaceMetadata(
+            isRepository: true,
+            branch: Self.gitBranchName(repository: repository),
+            isDirty: trackedChanges.isDirty,
+            indexSignature: trackedChanges.indexSignature,
+            indexContentSignature: trackedChanges.indexContentSignature,
+            headSignature: Self.gitHeadSignature(repository: repository)
+        )
+    }
+
+    /// The set of existing filesystem paths whose changes can alter the metadata
+    /// returned by ``workspaceMetadata(for:)`` for `directory`.
+    ///
+    /// Includes the working-tree root, `HEAD`, `index`, `refs`, `packed-refs`,
+    /// every reachable `config` (following `include`/`includeIf`), and the
+    /// equivalent paths for any gitlink submodules. Only paths that currently
+    /// exist are returned, sorted for stable comparison.
+    ///
+    /// - Parameter directory: An absolute path to inspect.
+    /// - Returns: Sorted existing paths to watch, or `nil` when `directory` is
+    ///   not inside a git repository.
+    public func watchedPaths(for directory: String) -> [String]? {
+        Self.workspaceGitMetadataWatchedPaths(for: directory)
+    }
+
+    /// The GitHub repository slugs (`owner/name`) configured as remotes for the
+    /// repository enclosing `directory`.
+    ///
+    /// Reads remote URLs straight from `config` (no `git` process), following
+    /// `include`/`includeIf`, and orders the result `upstream`, then `origin`,
+    /// then the rest, de-duplicated.
+    ///
+    /// - Parameter directory: An absolute path to inspect.
+    /// - Returns: Ordered, de-duplicated GitHub slugs; empty when there is no
+    ///   repository or no GitHub remote.
+    public func repositorySlugs(forDirectory directory: String) -> [String] {
+        guard let repository = Self.resolveGitRepository(containing: directory),
+              let output = Self.gitRemoteVOutput(repository: repository) else {
+            return []
+        }
+        return Self.githubRepositorySlugs(fromGitRemoteVOutput: output)
+    }
+}
