+++ Packages/CmuxGit/Tests/CmuxGitTests/GitMetadataServiceTests.swift
@@ -1,0 +1,235 @@
+import Foundation
+import Testing
+@testable import CmuxGit
+
+@Suite struct GitMetadataServiceTests {
+    // MARK: Repository resolution
+
+    @Test func resolvesNormalRepositoryFromNestedDirectory() throws {
+        let fixture = try GitRepositoryFixture()
+        try fixture.writeBranch("main")
+        let nested = fixture.root.appendingPathComponent("a/b/c")
+        try FileManager.default.createDirectory(at: nested, withIntermediateDirectories: true)
+
+        let repo = try #require(GitMetadataService.resolveGitRepository(containing: nested.path))
+        #expect(repo.workTreeRoot == fixture.root.standardizedFileURL.path)
+        #expect(repo.gitDirectory == fixture.gitDirectory.standardizedFileURL.path)
+        #expect(repo.commonDirectory == fixture.gitDirectory.standardizedFileURL.path)
+    }
+
+    @Test func returnsNilOutsideAnyRepository() {
+        let repo = GitMetadataService.resolveGitRepository(
+            containing: FileManager.default.temporaryDirectory.path
+        )
+        // The temp dir itself is not a repo; resolution must terminate at root.
+        #expect(repo == nil || repo?.workTreeRoot != FileManager.default.temporaryDirectory.path)
+    }
+
+    @Test func dotGitFilePointerResolvesGitDirectory() throws {
+        // Worktree-style: .git is a file pointing at a sibling git directory.
+        let base = FileManager.default.temporaryDirectory
+            .appendingPathComponent("cmuxgit-gitfile-\(UUID().uuidString)", isDirectory: true)
+        let worktree = base.appendingPathComponent("wt", isDirectory: true)
+        let realGitDir = base.appendingPathComponent("realgit", isDirectory: true)
+        try FileManager.default.createDirectory(at: worktree, withIntermediateDirectories: true)
+        try FileManager.default.createDirectory(at: realGitDir, withIntermediateDirectories: true)
+        defer { try? FileManager.default.removeItem(at: base) }
+
+        try "gitdir: \(realGitDir.path)\n".write(
+            to: worktree.appendingPathComponent(".git"),
+            atomically: true,
+            encoding: .utf8
+        )
+
+        let repo = try #require(GitMetadataService.resolveGitRepository(containing: worktree.path))
+        #expect(repo.workTreeRoot == worktree.standardizedFileURL.path)
+        #expect(repo.gitDirectory == realGitDir.standardizedFileURL.path)
+    }
+
+    @Test(arguments: [
+        ("/", "/"),
+        // Older Foundation can report /.. as the parent of /, or escape above
+        // root entirely; the search must still terminate in both cases.
+        ("/", "/.."),
+        ("/..", "/../.."),
+    ])
+    func rootVariantsStopRepositorySearch(current: String, parent: String) {
+        #expect(GitMetadataService.shouldStopGitRepositorySearch(
+            currentURL: URL(fileURLWithPath: current),
+            parentURL: URL(fileURLWithPath: parent)
+        ))
+    }
+
+    @Test func nonRootParentDoesNotStopSearch() {
+        #expect(!GitMetadataService.shouldStopGitRepositorySearch(
+            currentURL: URL(fileURLWithPath: "/Users/someone/project"),
+            parentURL: URL(fileURLWithPath: "/Users/someone")
+        ))
+    }
+
+    // MARK: Branch + metadata
+
+    @Test func workspaceMetadataReadsBranch() async throws {
+        let fixture = try GitRepositoryFixture()
+        try fixture.writeBranch("feature/x")
+        let service = GitMetadataService()
+        let meta = await service.workspaceMetadata(for: fixture.root.path)
+        #expect(meta.isRepository)
+        #expect(meta.branch == "feature/x")
+        #expect(meta.headSignature != nil)
+    }
+
+    @Test func workspaceMetadataReportsNotARepository() async {
+        let service = GitMetadataService()
+        let meta = await service.workspaceMetadata(for: "/definitely/not/a/repo/\(UUID().uuidString)")
+        #expect(meta == .notARepository)
+    }
+
+    @Test func detachedHeadHasNoBranch() async throws {
+        let fixture = try GitRepositoryFixture()
+        try fixture.writeDetachedHead(commit: String(repeating: "1", count: 40))
+        let service = GitMetadataService()
+        let meta = await service.workspaceMetadata(for: fixture.root.path)
+        #expect(meta.isRepository)
+        #expect(meta.branch == nil)
+    }
+
+    // MARK: Dirty detection (index v2)
+
+    @Test func cleanWorkingTreeIsNotDirty() async throws {
+        let fixture = try GitRepositoryFixture()
+        try fixture.writeBranch("main")
+        let entry = try fixture.writeWorkingTreeFile("file.txt", contents: "hello")
+        try fixture.writeIndex(GitIndexFixture(version: 2, entries: [entry]))
+
+        let service = GitMetadataService()
+        let meta = await service.workspaceMetadata(for: fixture.root.path)
+        #expect(meta.isRepository)
+        #expect(meta.isDirty == false)
+        #expect(meta.indexSignature != nil)
+        #expect(meta.indexContentSignature != nil)
+    }
+
+    @Test func modifiedSizeMarksDirty() async throws {
+        let fixture = try GitRepositoryFixture()
+        try fixture.writeBranch("main")
+        var entry = try fixture.writeWorkingTreeFile("file.txt", contents: "hello")
+        entry.size += 100 // index disagrees with the working tree
+        try fixture.writeIndex(GitIndexFixture(version: 2, entries: [entry]))
+
+        let service = GitMetadataService()
+        let meta = await service.workspaceMetadata(for: fixture.root.path)
+        #expect(meta.isDirty)
+    }
+
+    @Test func missingTrackedFileMarksDirty() async throws {
+        let fixture = try GitRepositoryFixture()
+        try fixture.writeBranch("main")
+        let entry = GitIndexFixture.Entry(path: "ghost.txt", size: 3)
+        try fixture.writeIndex(GitIndexFixture(version: 2, entries: [entry]))
+
+        let service = GitMetadataService()
+        let meta = await service.workspaceMetadata(for: fixture.root.path)
+        #expect(meta.isDirty)
+    }
+
+    @Test func assumeUnchangedEntryIsIgnoredForDirtiness() async throws {
+        let fixture = try GitRepositoryFixture()
+        try fixture.writeBranch("main")
+        var entry = try fixture.writeWorkingTreeFile("file.txt", contents: "hello")
+        entry.size += 100
+        entry.assumeUnchanged = true // excluded from the dirty check
+        try fixture.writeIndex(GitIndexFixture(version: 2, entries: [entry]))
+
+        let service = GitMetadataService()
+        let meta = await service.workspaceMetadata(for: fixture.root.path)
+        #expect(meta.isDirty == false)
+    }
+
+    // MARK: Index v4 prefix-compression
+
+    @Test func indexVersionFourDecodesPrefixCompressedPaths() throws {
+        let fixture = try GitRepositoryFixture()
+        try fixture.writeBranch("main")
+        let entries = [
+            GitIndexFixture.Entry(path: "src/alpha.swift"),
+            GitIndexFixture.Entry(path: "src/alphabet.swift"), // shares "src/alpha" prefix
+            GitIndexFixture.Entry(path: "src/beta.swift"),
+        ]
+        try fixture.writeIndex(GitIndexFixture(version: 4, entries: entries))
+
+        let indexURL = fixture.gitDirectory.appendingPathComponent("index")
+        let snapshot = try #require(GitMetadataService.gitIndexSnapshot(indexURL: indexURL))
+        #expect(snapshot.entries.map(\.path) == ["src/alpha.swift", "src/alphabet.swift", "src/beta.swift"])
+    }
+
+    @Test func indexVersionFourMultiByteStripLength() throws {
+        // A long shared prefix forces a multi-byte strip-length varint.
+        let longPrefix = String(repeating: "deep/", count: 40) // 200 chars
+        let fixture = try GitRepositoryFixture()
+        try fixture.writeBranch("main")
+        let entries = [
+            GitIndexFixture.Entry(path: longPrefix + "first.txt"),
+            GitIndexFixture.Entry(path: longPrefix + "second.txt"),
+        ]
+        try fixture.writeIndex(GitIndexFixture(version: 4, entries: entries))
+
+        let indexURL = fixture.gitDirectory.appendingPathComponent("index")
+        let snapshot = try #require(GitMetadataService.gitIndexSnapshot(indexURL: indexURL))
+        #expect(snapshot.entries.map(\.path) == [longPrefix + "first.txt", longPrefix + "second.txt"])
+    }
+
+    // MARK: Content signature stability
+
+    @Test func contentSignatureIgnoresStatOnlyChanges() {
+        let base = GitIndexEntryStat(
+            path: "a.txt", mode: 0o100644, objectID: String(repeating: "b", count: 40),
+            mtimeSeconds: 10, mtimeNanoseconds: 0, size: 5
+        )
+        let restated = GitIndexEntryStat(
+            path: "a.txt", mode: 0o100644, objectID: String(repeating: "b", count: 40),
+            mtimeSeconds: 999, mtimeNanoseconds: 7, size: 9999
+        )
+        let lhs = GitMetadataService.gitIndexContentSignature(entries: [base])
+        let rhs = GitMetadataService.gitIndexContentSignature(entries: [restated])
+        #expect(lhs == rhs) // path + mode + objectID only
+    }
+
+    @Test func contentSignatureChangesWithObjectID() {
+        let base = GitIndexEntryStat(
+            path: "a.txt", mode: 0o100644, objectID: String(repeating: "b", count: 40),
+            mtimeSeconds: 1, mtimeNanoseconds: 0, size: 0
+        )
+        let changed = GitIndexEntryStat(
+            path: "a.txt", mode: 0o100644, objectID: String(repeating: "c", count: 40),
+            mtimeSeconds: 1, mtimeNanoseconds: 0, size: 0
+        )
+        #expect(
+            GitMetadataService.gitIndexContentSignature(entries: [base])
+                != GitMetadataService.gitIndexContentSignature(entries: [changed])
+        )
+    }
+
+    // MARK: Watched paths
+
+    @Test func watchedPathsIncludeHeadIndexRefsAndConfig() async throws {
+        let fixture = try GitRepositoryFixture()
+        try fixture.writeBranch("main")
+        try fixture.writeConfig("[core]\n\trepositoryformatversion = 0\n")
+        try fixture.writeIndex(GitIndexFixture(version: 2, entries: []))
+
+        let service = GitMetadataService()
+        let paths = try #require(await service.watchedPaths(for: fixture.root.path))
+        #expect(paths == paths.sorted())
+        #expect(paths.contains(fixture.gitDirectory.appendingPathComponent("HEAD").standardizedFileURL.path))
+        #expect(paths.contains(fixture.gitDirectory.appendingPathComponent("index").standardizedFileURL.path))
+        #expect(paths.contains(fixture.gitDirectory.appendingPathComponent("config").standardizedFileURL.path))
+        #expect(paths.contains(fixture.root.standardizedFileURL.path))
+    }
+
+    @Test func watchedPathsNilOutsideRepository() async {
+        let service = GitMetadataService()
+        let paths = await service.watchedPaths(for: "/nope/\(UUID().uuidString)")
+        #expect(paths == nil)
+    }
+}
