+++ Sources/TabManager.swift
@@ -3,6 +3,7 @@ import SwiftUI
 import Foundation
 import Bonsplit
 import CmuxFileWatch
+import CmuxGit
 import CmuxProcess
 import CoreVideo
 import Combine
@@ -969,27 +970,6 @@ class TabManager: ObservableObject {
         let pullRequest: WorkspacePullRequestSnapshot
     }
 
-    private struct ResolvedGitRepository: Equatable, Sendable {
-        let workTreeRoot: String
-        let gitDirectory: String
-        let commonDirectory: String
-    }
-
-    private struct GitIndexEntryStat: Sendable {
-        let path: String
-        let mode: UInt32
-        let objectID: String
-        let mtimeSeconds: UInt32
-        let mtimeNanoseconds: UInt32
-        let size: UInt32
-    }
-
-    private struct GitIndexSnapshot: Sendable {
-        let entries: [GitIndexEntryStat]
-        let signature: String
-        let contentSignature: String
-    }
-
     private struct WorkspaceGitMetadataWatcherDescriptorRequest: Equatable, Sendable {
         let generation: UInt64
         let directory: String
@@ -1360,14 +1340,21 @@ class TabManager: ObservableObject {
     // tests can supply a fake without spawning a real process.
     private let commandRunner: any CommandRunning
 
+    // Reads on-disk git metadata (branch, dirty state, watched paths, remote
+    // slugs) off the main actor. Stateless; the reads are pure functions of the
+    // directory argument.
+    private let gitMetadataService: GitMetadataService
+
     init(
         initialWorkspaceTitle: String? = nil,
         initialWorkingDirectory: String? = nil,
         initialTerminalInput: String? = nil,
         autoWelcomeIfNeeded: Bool = true,
-        commandRunner: any CommandRunning = CommandRunner()
+        commandRunner: any CommandRunning = CommandRunner(),
+        gitMetadataService: GitMetadataService = GitMetadataService()
     ) {
         self.commandRunner = commandRunner
+        self.gitMetadataService = gitMetadataService
         addWorkspace(
             title: initialWorkspaceTitle,
             workingDirectory: initialWorkingDirectory,
@@ -1652,9 +1639,8 @@ class TabManager: ObservableObject {
         workspaceGitMetadataWatcherDescriptorRequestsByKey[key] = request
 
         Task { [weak self] in
-            let watchedPaths = await Task.detached(priority: .utility) {
-                Self.workspaceGitMetadataWatchedPaths(for: directory)
-            }.value
+            guard let gitMetadataService = self?.gitMetadataService else { return }
+            let watchedPaths = await gitMetadataService.watchedPaths(for: directory)
             await MainActor.run { [weak self] in
                 self?.applyWorkspaceGitMetadataWatcherDescriptor(
                     watchedPaths,
@@ -1814,8 +1800,12 @@ class TabManager: ObservableObject {
         let allowCachedResults = allowCachedResultsOverride
             ?? Self.workspacePullRequestRefreshAllowsRepoCache(reason: reason)
         let commandRunner = commandRunner
+        let gitMetadataService = gitMetadataService
         workspacePullRequestRefreshTask = Task.detached(priority: .utility) { [weak self] in
-            let candidateResolution = await Self.resolveWorkspacePullRequestCandidateSeeds(candidateSeeds)
+            let candidateResolution = await Self.resolveWorkspacePullRequestCandidateSeeds(
+                candidateSeeds,
+                gitMetadataService: gitMetadataService
+            )
             guard !Task.isCancelled else { return }
             let repoResults = await Self.fetchWorkspacePullRequestRepoResults(
                 repoDirectoriesBySlug: candidateResolution.repoDirectoriesBySlug,
@@ -1876,7 +1866,8 @@ class TabManager: ObservableObject {
     @concurrent
     #endif
     private nonisolated static func resolveWorkspacePullRequestCandidateSeeds(
-        _ seeds: [WorkspacePullRequestCandidateSeed]
+        _ seeds: [WorkspacePullRequestCandidateSeed],
+        gitMetadataService: GitMetadataService
     ) async -> WorkspacePullRequestCandidateResolution {
         var candidates: [WorkspacePullRequestCandidate] = []
         candidates.reserveCapacity(seeds.count)
@@ -1890,7 +1881,7 @@ class TabManager: ObservableObject {
                 if let cachedRepoSlugs = repoSlugsByDirectory[directory] {
                     repoSlugs = cachedRepoSlugs
                 } else {
-                    let resolvedRepoSlugs = await githubRepositorySlugs(directory: directory)
+                    let resolvedRepoSlugs = await gitMetadataService.repositorySlugs(forDirectory: directory)
                     repoSlugsByDirectory[directory] = resolvedRepoSlugs
                     repoSlugs = resolvedRepoSlugs
                 }
@@ -2941,7 +2932,9 @@ class TabManager: ObservableObject {
         }
 
         Task.detached(priority: .utility) { [weak self] in
-            let snapshot = await Self.initialWorkspaceGitMetadataSnapshot(for: expectedDirectory)
+            guard let snapshot = await self?.initialWorkspaceGitMetadataSnapshot(for: expectedDirectory) else {
+                return
+            }
             guard !Task.isCancelled else { return }
             await MainActor.run { [weak self] in
                 self?.applyWorkspaceGitMetadataSnapshot(
@@ -3231,10 +3224,11 @@ class TabManager: ObservableObject {
         }
     }
 
-    private nonisolated static func initialWorkspaceGitMetadataSnapshot(
+    private nonisolated func initialWorkspaceGitMetadataSnapshot(
         for directory: String
     ) async -> InitialWorkspaceGitMetadataSnapshot {
-        guard let repository = resolveGitRepository(containing: directory) else {
+        let metadata = await gitMetadataService.workspaceMetadata(for: directory)
+        guard metadata.isRepository else {
             return InitialWorkspaceGitMetadataSnapshot(
                 isRepository: false,
                 branch: nil,
@@ -3245,1006 +3239,17 @@ class TabManager: ObservableObject {
                 pullRequest: .notFound
             )
         }
-        let trackedChanges = gitTrackedChangesSnapshot(repository: repository)
-        let headSignature = gitHeadSignature(repository: repository)
 
-        let branch = normalizedBranchName(gitBranchName(repository: repository))
-        guard let branch else {
-            return InitialWorkspaceGitMetadataSnapshot(
-                isRepository: true,
-                branch: nil,
-                isDirty: trackedChanges.isDirty,
-                indexSignature: trackedChanges.indexSignature,
-                indexContentSignature: trackedChanges.indexContentSignature,
-                headSignature: headSignature,
-                pullRequest: .notFound
-            )
-        }
-
+        let branch = Self.normalizedBranchName(metadata.branch)
         return InitialWorkspaceGitMetadataSnapshot(
             isRepository: true,
             branch: branch,
-            isDirty: trackedChanges.isDirty,
-            indexSignature: trackedChanges.indexSignature,
-            indexContentSignature: trackedChanges.indexContentSignature,
-            headSignature: headSignature,
-            pullRequest: .deferred
+            isDirty: metadata.isDirty,
+            indexSignature: metadata.indexSignature,
+            indexContentSignature: metadata.indexContentSignature,
+            headSignature: metadata.headSignature,
+            pullRequest: branch == nil ? .notFound : .deferred
         )
-    }
-
-    private nonisolated static func resolveGitRepository(containing directory: String) -> ResolvedGitRepository? {
-        let startURL = URL(fileURLWithPath: directory).standardizedFileURL
-        let fileManager = FileManager.default
-        var currentURL = startURL
-        var isDirectory: ObjCBool = false
-
-        if !fileManager.fileExists(atPath: currentURL.path, isDirectory: &isDirectory) || !isDirectory.boolValue {
-            currentURL.deleteLastPathComponent()
-        }
-
-        while true {
-            let dotGitURL = currentURL.appendingPathComponent(".git")
-            if fileManager.fileExists(atPath: dotGitURL.path, isDirectory: &isDirectory) {
-                let gitDirectory: String?
-                if isDirectory.boolValue {
-                    gitDirectory = dotGitURL.standardizedFileURL.path
-                } else {
-                    gitDirectory = gitDirectoryFromDotGitFile(dotGitURL, relativeTo: currentURL)
-                }
-
-                if let gitDirectory {
-                    let commonDirectory = gitCommonDirectory(gitDirectory: gitDirectory)
-                    return ResolvedGitRepository(
-                        workTreeRoot: currentURL.standardizedFileURL.path,
-                        gitDirectory: gitDirectory,
-                        commonDirectory: commonDirectory
-                    )
-                }
-            }
-
-            let parentURL = currentURL.deletingLastPathComponent()
-            if Self.shouldStopGitRepositorySearch(currentURL: currentURL, parentURL: parentURL) {
-                return nil
-            }
-            currentURL = parentURL
-        }
-    }
-
-    nonisolated static func shouldStopGitRepositorySearch(currentURL: URL, parentURL: URL) -> Bool {
-        if parentURL.path == currentURL.path {
-            return true
-        }
-
-        let standardizedCurrentPath = currentURL.standardizedFileURL.path
-        if standardizedCurrentPath == "/" {
-            return true
-        }
-
-        return parentURL.standardizedFileURL.path == standardizedCurrentPath
-    }
-
-    private nonisolated static func gitDirectoryFromDotGitFile(
-        _ dotGitURL: URL,
-        relativeTo workTreeRootURL: URL
-    ) -> String? {
-        guard let contents = try? String(contentsOf: dotGitURL, encoding: .utf8) else {
-            return nil
-        }
-        let trimmed = contents.trimmingCharacters(in: .whitespacesAndNewlines)
-        let prefix = "gitdir:"
-        guard trimmed.lowercased().hasPrefix(prefix) else {
-            return nil
-        }
-
-        let rawPath = trimmed.dropFirst(prefix.count).trimmingCharacters(in: .whitespacesAndNewlines)
-        guard !rawPath.isEmpty else { return nil }
-        if rawPath.hasPrefix("/") {
-            return URL(fileURLWithPath: String(rawPath)).standardizedFileURL.path
-        }
-        return workTreeRootURL
-            .appendingPathComponent(String(rawPath))
-            .standardizedFileURL
-            .path
-    }
-
-    private nonisolated static func gitCommonDirectory(gitDirectory: String) -> String {
-        let gitDirectoryURL = URL(fileURLWithPath: gitDirectory)
-        let commonDirURL = gitDirectoryURL.appendingPathComponent("commondir")
-        guard let contents = try? String(contentsOf: commonDirURL, encoding: .utf8) else {
-            return gitDirectory
-        }
-
-        let rawPath = contents.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard !rawPath.isEmpty else { return gitDirectory }
-        if rawPath.hasPrefix("/") {
-            return URL(fileURLWithPath: rawPath).standardizedFileURL.path
-        }
-        return gitDirectoryURL
-            .appendingPathComponent(rawPath)
-            .standardizedFileURL
-            .path
-    }
-
-    private nonisolated static func workspaceGitMetadataWatchedPaths(
-        for directory: String
-    ) -> [String]? {
-        guard let repository = resolveGitRepository(containing: directory) else {
-            return nil
-        }
-
-        let candidatePaths = [
-            repository.workTreeRoot,
-        ] + gitRepositoryMetadataWatchPaths(repository: repository)
-            + gitlinkMetadataWatchPaths(repository: repository)
-        var watchedPaths: [String] = []
-        var seen: Set<String> = []
-        for path in candidatePaths {
-            let normalized = URL(fileURLWithPath: path).standardizedFileURL.path
-            guard seen.insert(normalized).inserted else { continue }
-            var isDirectory: ObjCBool = false
-            guard FileManager.default.fileExists(atPath: normalized, isDirectory: &isDirectory) else {
-                continue
-            }
-            watchedPaths.append(normalized)
-        }
-
-        return watchedPaths.sorted()
-    }
-
-    private nonisolated static func gitRepositoryMetadataWatchPaths(
-        repository: ResolvedGitRepository
-    ) -> [String] {
-        [
-            URL(fileURLWithPath: repository.gitDirectory).appendingPathComponent("HEAD").path,
-            URL(fileURLWithPath: repository.gitDirectory).appendingPathComponent("index").path,
-            URL(fileURLWithPath: repository.gitDirectory).appendingPathComponent("refs").path,
-            URL(fileURLWithPath: repository.commonDirectory).appendingPathComponent("refs").path,
-            URL(fileURLWithPath: repository.commonDirectory).appendingPathComponent("packed-refs").path,
-        ] + gitConfigURLs(repository: repository).map(\.path)
-    }
-
-    private nonisolated static func gitlinkMetadataWatchPaths(
-        repository: ResolvedGitRepository
-    ) -> [String] {
-        let indexURL = URL(fileURLWithPath: repository.gitDirectory).appendingPathComponent("index")
-        guard let indexSnapshot = gitIndexSnapshot(indexURL: indexURL) else {
-            return []
-        }
-
-        let gitlinkMode: UInt32 = 0o160000
-        var paths: [String] = []
-        for entry in indexSnapshot.entries where (entry.mode & 0o170000) == gitlinkMode {
-            let gitlinkURL = URL(fileURLWithPath: repository.workTreeRoot)
-                .appendingPathComponent(entry.path)
-                .standardizedFileURL
-            guard let submoduleRepository = resolveGitRepository(containing: gitlinkURL.path),
-                  submoduleRepository.workTreeRoot == gitlinkURL.path else {
-                continue
-            }
-            paths.append(contentsOf: gitRepositoryMetadataWatchPaths(repository: submoduleRepository))
-        }
-        return paths
-    }
-
-    private nonisolated static func gitBranchName(repository: ResolvedGitRepository) -> String? {
-        let headURL = URL(fileURLWithPath: repository.gitDirectory).appendingPathComponent("HEAD")
-        guard let contents = try? String(contentsOf: headURL, encoding: .utf8) else {
-            return nil
-        }
-        let trimmed = contents.trimmingCharacters(in: .whitespacesAndNewlines)
-        let branchPrefix = "ref: refs/heads/"
-        guard trimmed.hasPrefix(branchPrefix) else {
-            return nil
-        }
-        let branch = String(trimmed.dropFirst(branchPrefix.count))
-        return branch.isEmpty ? nil : branch
-    }
-
-    private nonisolated static func gitHeadSignature(repository: ResolvedGitRepository) -> String? {
-        let headURL = URL(fileURLWithPath: repository.gitDirectory).appendingPathComponent("HEAD")
-        guard let contents = try? String(contentsOf: headURL, encoding: .utf8) else {
-            return nil
-        }
-        let trimmed = contents.trimmingCharacters(in: .whitespacesAndNewlines)
-        let refPrefix = "ref: "
-        guard trimmed.hasPrefix(refPrefix) else {
-            return trimmed.isEmpty ? nil : trimmed
-        }
-
-        let refName = String(trimmed.dropFirst(refPrefix.count))
-        guard !refName.isEmpty else { return trimmed }
-        let refValue = gitRefValue(repository: repository, refName: refName) ?? ""
-        return "\(trimmed)\n\(refValue)"
-    }
-
-    private nonisolated static func gitRefValue(repository: ResolvedGitRepository, refName: String) -> String? {
-        let refURLs = [
-            URL(fileURLWithPath: repository.gitDirectory).appendingPathComponent(refName),
-            URL(fileURLWithPath: repository.commonDirectory).appendingPathComponent(refName),
-        ]
-        var seenPaths: Set<String> = []
-        for refURL in refURLs {
-            let path = refURL.standardizedFileURL.path
-            guard seenPaths.insert(path).inserted,
-                  let contents = try? String(contentsOf: refURL, encoding: .utf8) else {
-                continue
-            }
-            let value = contents.trimmingCharacters(in: .whitespacesAndNewlines)
-            if !value.isEmpty {
-                return value
-            }
-        }
-
-        let packedRefsURL = URL(fileURLWithPath: repository.commonDirectory).appendingPathComponent("packed-refs")
-        guard let packedRefs = try? String(contentsOf: packedRefsURL, encoding: .utf8) else {
-            return nil
-        }
-        for rawLine in packedRefs.components(separatedBy: .newlines) {
-            let line = rawLine.trimmingCharacters(in: .whitespaces)
-            guard !line.isEmpty, !line.hasPrefix("#"), !line.hasPrefix("^") else { continue }
-            let parts = line.split(whereSeparator: { $0 == " " || $0 == "\t" })
-            guard parts.count == 2, String(parts[1]) == refName else { continue }
-            return String(parts[0])
-        }
-        return nil
-    }
-
-    private nonisolated static func gitRemoteVOutput(repository: ResolvedGitRepository) -> String? {
-        var lines: [String] = []
-        var seenConfigPaths: Set<String> = []
-        for configURL in gitRootConfigURLs(repository: repository) {
-            appendGitRemoteVLines(
-                fromConfigURL: configURL,
-                repository: repository,
-                seenConfigPaths: &seenConfigPaths,
-                lines: &lines
-            )
-        }
-        return lines.isEmpty ? nil : lines.joined()
-    }
-
-    private nonisolated static func gitRootConfigURLs(repository: ResolvedGitRepository) -> [URL] {
-        [
-            URL(fileURLWithPath: repository.commonDirectory).appendingPathComponent("config"),
-            URL(fileURLWithPath: repository.gitDirectory).appendingPathComponent("config"),
-        ]
-    }
-
-    private nonisolated static func gitConfigURLs(repository: ResolvedGitRepository) -> [URL] {
-        var urls: [URL] = []
-        var pendingURLs = gitRootConfigURLs(repository: repository)
-        var seenConfigPaths: Set<String> = []
-
-        while !pendingURLs.isEmpty {
-            let configURL = pendingURLs.removeFirst().standardizedFileURL
-            let path = configURL.path
-            guard seenConfigPaths.insert(path).inserted else { continue }
-            urls.append(configURL)
-            guard let config = try? String(contentsOf: configURL, encoding: .utf8) else {
-                continue
-            }
-            pendingURLs.append(
-                contentsOf: gitIncludedConfigURLs(
-                    fromConfig: config,
-                    configURL: configURL,
-                    repository: repository
-                )
-            )
-        }
-
-        return urls
-    }
-
-    private nonisolated static func gitRemoteVLines(fromConfig config: String) -> [String] {
-        var currentRemoteName: String?
-        var lines: [String] = []
-
-        for rawLine in config.components(separatedBy: .newlines) {
-            let line = gitConfigLineRemovingInlineComment(rawLine)
-                .trimmingCharacters(in: .whitespaces)
-            if line.hasPrefix("[") && line.hasSuffix("]") {
-                currentRemoteName = gitConfigRemoteName(fromSectionHeader: line)
-                continue
-            }
-
-            guard let currentRemoteName else { continue }
-            let parts = line.split(separator: "=", maxSplits: 1).map {
-                $0.trimmingCharacters(in: .whitespaces)
-            }
-            guard parts.count == 2, parts[0] == "url" else {
-                continue
-            }
-            let remoteURL = gitConfigUnquotedValue(parts[1])
-            guard !remoteURL.isEmpty else {
-                continue
-            }
-            lines.append("\(currentRemoteName)\t\(remoteURL) (fetch)\n")
-        }
-
-        return lines
-    }
-
-    private nonisolated static func appendGitRemoteVLines(
-        fromConfigURL configURL: URL,
-        repository: ResolvedGitRepository,
-        seenConfigPaths: inout Set<String>,
-        lines: inout [String]
-    ) {
-        let configURL = configURL.standardizedFileURL
-        guard seenConfigPaths.insert(configURL.path).inserted,
-              let config = try? String(contentsOf: configURL, encoding: .utf8) else {
-            return
-        }
-
-        var currentRemoteName: String?
-        var currentSectionAllowsIncludePath = false
-
-        for rawLine in config.components(separatedBy: .newlines) {
-            let line = gitConfigLineRemovingInlineComment(rawLine)
-                .trimmingCharacters(in: .whitespaces)
-            if line.hasPrefix("[") && line.hasSuffix("]") {
-                currentRemoteName = gitConfigRemoteName(fromSectionHeader: line)
-                if line == "[include]" {
-                    currentSectionAllowsIncludePath = true
-                } else if let condition = gitConfigIncludeIfCondition(fromSectionHeader: line) {
-                    currentSectionAllowsIncludePath = gitConfigIncludeIfConditionMatches(
-                        condition,
-                        repository: repository
-                    )
-                } else {
-                    currentSectionAllowsIncludePath = false
-                }
-                continue
-            }
-
-            let parts = line.split(separator: "=", maxSplits: 1).map {
-                $0.trimmingCharacters(in: .whitespaces)
-            }
-
-            if let currentRemoteName,
-               parts.count == 2,
-               parts[0] == "url" {
-                let remoteURL = gitConfigUnquotedValue(parts[1])
-                guard !remoteURL.isEmpty else {
-                    continue
-                }
-                lines.append("\(currentRemoteName)\t\(remoteURL) (fetch)\n")
-                continue
-            }
-
-            guard currentSectionAllowsIncludePath,
-                  parts.count == 2,
-                  parts[0] == "path",
-                  let includeURL = gitConfigIncludeURL(
-                      fromPathValue: parts[1],
-                      relativeTo: configURL
-                  ) else {
-                continue
-            }
-            appendGitRemoteVLines(
-                fromConfigURL: includeURL,
-                repository: repository,
-                seenConfigPaths: &seenConfigPaths,
-                lines: &lines
-            )
-        }
-    }
-
-    private nonisolated static func gitIncludedConfigURLs(
-        fromConfig config: String,
-        configURL: URL,
-        repository: ResolvedGitRepository
-    ) -> [URL] {
-        var currentSectionAllowsPath = false
-        var urls: [URL] = []
-
-        for rawLine in config.components(separatedBy: .newlines) {
-            let line = gitConfigLineRemovingInlineComment(rawLine)
-                .trimmingCharacters(in: .whitespaces)
-            if line.hasPrefix("[") && line.hasSuffix("]") {
-                if line == "[include]" {
-                    currentSectionAllowsPath = true
-                } else if let condition = gitConfigIncludeIfCondition(fromSectionHeader: line) {
-                    currentSectionAllowsPath = gitConfigIncludeIfConditionMatches(
-                        condition,
-                        repository: repository
-                    )
-                } else {
-                    currentSectionAllowsPath = false
-                }
-                continue
-            }
-
-            guard currentSectionAllowsPath else { continue }
-            let parts = line.split(separator: "=", maxSplits: 1).map {
-                $0.trimmingCharacters(in: .whitespaces)
-            }
-            guard parts.count == 2,
-                  parts[0].lowercased() == "path",
-                  let includeURL = gitConfigIncludeURL(
-                    fromPathValue: parts[1],
-                    relativeTo: configURL
-                  ) else {
-                continue
-            }
-            urls.append(includeURL)
-        }
-
-        return urls
-    }
-
-    private nonisolated static func gitConfigUnquotedValue(_ value: String) -> String {
-        let trimmedValue = value.trimmingCharacters(in: .whitespaces)
-        guard trimmedValue.first == "\"",
-              trimmedValue.last == "\"",
-              trimmedValue.count >= 2 else {
-            return trimmedValue
-        }
-
-        var result = ""
-        var isEscaped = false
-        for character in trimmedValue.dropFirst().dropLast() {
-            if isEscaped {
-                result.append(character)
-                isEscaped = false
-                continue
-            }
-
-            if character == "\\" {
-                isEscaped = true
-                continue
-            }
-
-            result.append(character)
-        }
-
-        if isEscaped {
-            result.append("\\")
-        }
-        return result
-    }
-
-    private nonisolated static func gitConfigLineRemovingInlineComment(_ line: String) -> String {
-        var result = ""
-        var isInsideDoubleQuotedString = false
-        var isEscaped = false
-        var previousWasWhitespace = true
-
-        for character in line {
-            if isEscaped {
-                result.append(character)
-                isEscaped = false
-                previousWasWhitespace = character.isWhitespace
-                continue
-            }
-
-            if isInsideDoubleQuotedString && character == "\\" {
-                result.append(character)
-                isEscaped = true
-                previousWasWhitespace = false
-                continue
-            }
-
-            if character == "\"" {
-                result.append(character)
-                isInsideDoubleQuotedString.toggle()
-                previousWasWhitespace = false
-                continue
-            }
-
-            if !isInsideDoubleQuotedString,
-               previousWasWhitespace,
-               (character == "#" || character == ";") {
-                break
-            }
-
-            result.append(character)
-            previousWasWhitespace = character.isWhitespace
-        }
-
-        return result
-    }
-
-    private nonisolated static func gitConfigRemoteName(fromSectionHeader header: String) -> String? {
-        let prefix = "[remote \""
-        let suffix = "\"]"
-        guard header.hasPrefix(prefix), header.hasSuffix(suffix) else {
-            return nil
-        }
-        let name = header.dropFirst(prefix.count).dropLast(suffix.count)
-        return name.isEmpty ? nil : String(name)
-    }
-
-    private nonisolated static func gitConfigIncludeIfCondition(fromSectionHeader header: String) -> String? {
-        let prefix = "[includeIf \""
-        let suffix = "\"]"
-        guard header.hasPrefix(prefix), header.hasSuffix(suffix) else {
-            return nil
-        }
-        let condition = header.dropFirst(prefix.count).dropLast(suffix.count)
-        return condition.isEmpty ? nil : String(condition)
-    }
-
-    private nonisolated static func gitConfigIncludeURL(
-        fromPathValue pathValue: String,
-        relativeTo configURL: URL
-    ) -> URL? {
-        let path = gitConfigUnquotedValue(pathValue)
-        guard !path.isEmpty else { return nil }
-        if path == "~" {
-            return FileManager.default.homeDirectoryForCurrentUser.standardizedFileURL
-        }
-        if path.hasPrefix("~/") {
-            let relativePath = String(path.dropFirst(2))
-            return FileManager.default.homeDirectoryForCurrentUser
-                .appendingPathComponent(relativePath)
-                .standardizedFileURL
-        }
-        if path.hasPrefix("/") {
-            return URL(fileURLWithPath: path).standardizedFileURL
-        }
-        return configURL
-            .deletingLastPathComponent()
-            .appendingPathComponent(path)
-            .standardizedFileURL
-    }
-
-    private nonisolated static func gitConfigIncludeIfConditionMatches(
-        _ condition: String,
-        repository: ResolvedGitRepository
-    ) -> Bool {
-        let lowercasedCondition = condition.lowercased()
-        if lowercasedCondition.hasPrefix("gitdir/i:") {
-            let pattern = String(condition.dropFirst("gitdir/i:".count))
-            return gitConfigGitdirPatternMatches(pattern, repository: repository, caseInsensitive: true)
-        }
-        if lowercasedCondition.hasPrefix("gitdir:") {
-            let pattern = String(condition.dropFirst("gitdir:".count))
-            return gitConfigGitdirPatternMatches(pattern, repository: repository, caseInsensitive: false)
-        }
-        if lowercasedCondition.hasPrefix("onbranch:") {
-            let pattern = String(condition.dropFirst("onbranch:".count))
-            guard let branch = gitBranchName(repository: repository) else { return false }
-            return gitConfigGlobMatches(branch, pattern: pattern, caseInsensitive: false)
-        }
-        return false
-    }
-
-    private nonisolated static func gitConfigGitdirPatternMatches(
-        _ pattern: String,
-        repository: ResolvedGitRepository,
-        caseInsensitive: Bool
-    ) -> Bool {
-        let isRecursiveDirectoryPattern = pattern.hasSuffix("/")
-        var expandedPattern = gitConfigExpandedPattern(pattern)
-        if isRecursiveDirectoryPattern, !expandedPattern.hasSuffix("/") {
-            expandedPattern.append("/")
-        }
-        if isRecursiveDirectoryPattern {
-            expandedPattern.append("**")
-        }
-        let candidates = [
-            repository.gitDirectory,
-            repository.commonDirectory,
-            repository.workTreeRoot,
-        ].map { URL(fileURLWithPath: $0).standardizedFileURL.path }
-
-        for candidate in candidates {
-            if gitConfigGlobMatches(candidate, pattern: expandedPattern, caseInsensitive: caseInsensitive) ||
-                gitConfigGlobMatches(candidate + "/", pattern: expandedPattern, caseInsensitive: caseInsensitive) {
-                return true
-            }
-        }
-        return false
-    }
-
-    private nonisolated static func gitConfigExpandedPattern(_ pattern: String) -> String {
-        if pattern == "~" {
-            return FileManager.default.homeDirectoryForCurrentUser.standardizedFileURL.path
-        }
-        if pattern.hasPrefix("~/") {
-            let relativePath = String(pattern.dropFirst(2))
-            return FileManager.default.homeDirectoryForCurrentUser
-                .appendingPathComponent(relativePath)
-                .standardizedFileURL
-                .path
-        }
-        return URL(fileURLWithPath: pattern).standardizedFileURL.path
-    }
-
-    private nonisolated static func gitConfigGlobMatches(
-        _ value: String,
-        pattern: String,
-        caseInsensitive: Bool
-    ) -> Bool {
-        let candidateValue = caseInsensitive ? value.lowercased() : value
-        let candidatePattern = caseInsensitive ? pattern.lowercased() : pattern
-        guard let regex = try? NSRegularExpression(
-            pattern: gitConfigGlobRegexPattern(candidatePattern)
-        ) else {
-            return fnmatch(candidatePattern, candidateValue, 0) == 0
-        }
-        let range = NSRange(candidateValue.startIndex..<candidateValue.endIndex, in: candidateValue)
-        return regex.firstMatch(in: candidateValue, range: range) != nil
-    }
-
-    private nonisolated static func gitConfigGlobRegexPattern(_ pattern: String) -> String {
-        let characters = Array(pattern)
-        var regex = "^"
-        var index = 0
-
-        while index < characters.count {
-            let character = characters[index]
-
-            if character == "*" {
-                var starCount = 1
-                while index + starCount < characters.count,
-                      characters[index + starCount] == "*" {
-                    starCount += 1
-                }
-                index += starCount
-
-                if starCount >= 2 {
-                    if index < characters.count, characters[index] == "/" {
-                        index += 1
-                        regex += "(?:.*/)?"
-                    } else {
-                        regex += ".*"
-                    }
-                } else {
-                    regex += "[^/]*"
-                }
-                continue
-            }
-
-            if character == "?" {
-                regex += "[^/]"
-                index += 1
-                continue
-            }
-
-            if character == "[" {
-                let parsedClass = gitConfigGlobCharacterClass(characters, startIndex: index)
-                if let parsedClass {
-                    regex += parsedClass.regex
-                    index = parsedClass.endIndex
-                    continue
-                }
-            }
-
-            regex += NSRegularExpression.escapedPattern(for: String(character))
-            index += 1
-        }
-
-        regex += "$"
-        return regex
-    }
-
-    private nonisolated static func gitConfigGlobCharacterClass(
-        _ characters: [Character],
-        startIndex: Int
-    ) -> (regex: String, endIndex: Int)? {
-        guard startIndex < characters.count, characters[startIndex] == "[" else {
-            return nil
-        }
-
-        var index = startIndex + 1
-        guard index < characters.count else { return nil }
-
-        var regex = "["
-        if characters[index] == "!" {
-            regex += "^"
-            index += 1
-        } else if characters[index] == "^" {
-            regex += "\\^"
-            index += 1
-        }
-
-        if index < characters.count, characters[index] == "]" {
-            regex += "\\]"
-            index += 1
-        }
-
-        var hasTerminator = false
-        while index < characters.count {
-            let character = characters[index]
-            if character == "]" {
-                hasTerminator = true
-                index += 1
-                break
-            }
-            switch character {
-            case "\\":
-                regex += "\\\\"
-            case "[":
-                regex += "\\["
-            default:
-                regex += String(character)
-            }
-            index += 1
-        }
-
-        guard hasTerminator else { return nil }
-        regex += "]"
-        return (regex, index)
-    }
-
-    private nonisolated static func gitTrackedChangesSnapshot(
-        repository: ResolvedGitRepository
-    ) -> (isDirty: Bool, indexSignature: String?, indexContentSignature: String?) {
-        let indexURL = URL(fileURLWithPath: repository.gitDirectory).appendingPathComponent("index")
-        guard let indexSnapshot = gitIndexSnapshot(indexURL: indexURL) else {
-            return (false, gitIndexFileSignature(indexURL: indexURL), nil)
-        }
-
-        for entry in indexSnapshot.entries {
-            let fileURL = URL(fileURLWithPath: repository.workTreeRoot).appendingPathComponent(entry.path)
-            let gitlinkMode: UInt32 = 0o160000
-            if (entry.mode & 0o170000) == gitlinkMode {
-                guard let submoduleCommit = gitlinkWorktreeCommit(
-                    parentRepository: repository,
-                    gitlinkPath: entry.path
-                ) else {
-                    return (true, indexSnapshot.signature, indexSnapshot.contentSignature)
-                }
-                if submoduleCommit.caseInsensitiveCompare(entry.objectID) != .orderedSame {
-                    return (true, indexSnapshot.signature, indexSnapshot.contentSignature)
-                }
-                continue
-            }
-
-            var statValue = stat()
-            guard lstat(fileURL.path, &statValue) == 0 else {
-                return (true, indexSnapshot.signature, indexSnapshot.contentSignature)
-            }
-            let size = gitIndexUInt32Field(statValue.st_size)
-            let mtimeSeconds = gitIndexUInt32Field(statValue.st_mtimespec.tv_sec)
-            let mtimeNanoseconds = gitIndexUInt32Field(statValue.st_mtimespec.tv_nsec)
-            guard let mode = gitIndexComparableMode(for: statValue.st_mode) else {
-                return (true, indexSnapshot.signature, indexSnapshot.contentSignature)
-            }
-            if size != entry.size ||
-                mode != entry.mode ||
-                mtimeSeconds != entry.mtimeSeconds ||
-                mtimeNanoseconds != entry.mtimeNanoseconds {
-                return (true, indexSnapshot.signature, indexSnapshot.contentSignature)
-            }
-        }
-
-        return (false, indexSnapshot.signature, indexSnapshot.contentSignature)
-    }
-
-    private nonisolated static func gitIndexSnapshot(indexURL: URL) -> GitIndexSnapshot? {
-        guard let data = try? Data(contentsOf: indexURL), data.count >= 32 else {
-            return nil
-        }
-        let bytes = [UInt8](data)
-        guard bytes[0] == 0x44, bytes[1] == 0x49, bytes[2] == 0x52, bytes[3] == 0x43 else {
-            return nil
-        }
-        let version = readBigEndianUInt32(bytes, at: 4)
-        guard version == 2 || version == 3 || version == 4 else {
-            return nil
-        }
-        let entryCount = Int(readBigEndianUInt32(bytes, at: 8))
-        let contentEnd = bytes.count - 20
-        var offset = 12
-        var entries: [GitIndexEntryStat] = []
-        var contentEntries: [GitIndexEntryStat] = []
-        entries.reserveCapacity(min(entryCount, 1024))
-        contentEntries.reserveCapacity(min(entryCount, 1024))
-        var previousPathBytes: [UInt8] = []
-
-        for _ in 0..<entryCount {
-            guard offset + 62 <= contentEnd else { return nil }
-            let entryStart = offset
-            let mtimeSeconds = readBigEndianUInt32(bytes, at: offset + 8)
-            let mtimeNanoseconds = readBigEndianUInt32(bytes, at: offset + 12)
-            let mode = readBigEndianUInt32(bytes, at: offset + 24)
-            let size = readBigEndianUInt32(bytes, at: offset + 36)
-            let objectID = bytes[(offset + 40)..<(offset + 60)].map {
-                String(format: "%02x", $0)
-            }.joined()
-            let flags = readBigEndianUInt16(bytes, at: offset + 60)
-            let pathLength = Int(flags & 0x0fff)
-            let hasExtendedFlags = version >= 3 && (flags & 0x4000) != 0
-            var extendedFlags: UInt16 = 0
-            offset += 62
-            if hasExtendedFlags {
-                guard offset + 2 <= contentEnd else { return nil }
-                extendedFlags = readBigEndianUInt16(bytes, at: offset)
-                offset += 2
-            }
-
-            let pathBytes: [UInt8]
-            if version == 4 {
-                guard let stripLength = readGitIndexV4PathStripLength(bytes, offset: &offset),
-                      stripLength <= previousPathBytes.count else {
-                    return nil
-                }
-                let suffixStart = offset
-                while offset < contentEnd, bytes[offset] != 0 {
-                    offset += 1
-                }
-                guard offset < contentEnd else { return nil }
-                pathBytes = Array(previousPathBytes.dropLast(stripLength)) + Array(bytes[suffixStart..<offset])
-            } else {
-                let pathStart = offset
-                if pathLength < 0x0fff {
-                    offset += pathLength
-                    guard offset < contentEnd else { return nil }
-                } else {
-                    while offset < contentEnd, bytes[offset] != 0 {
-                        offset += 1
-                    }
-                    guard offset < contentEnd else { return nil }
-                }
-                pathBytes = Array(bytes[pathStart..<offset])
-            }
-
-            let pathData = Data(pathBytes)
-            guard let path = String(data: pathData, encoding: .utf8), !path.isEmpty else {
-                return nil
-            }
-            previousPathBytes = pathBytes
-            let entryStat = GitIndexEntryStat(
-                path: path,
-                mode: mode,
-                objectID: objectID,
-                mtimeSeconds: mtimeSeconds,
-                mtimeNanoseconds: mtimeNanoseconds,
-                size: size
-            )
-            contentEntries.append(entryStat)
-
-            let assumeUnchangedFlag: UInt16 = 0x8000
-            let skipWorktreeExtendedFlag: UInt16 = 0x4000
-            if (flags & assumeUnchangedFlag) == 0,
-               (extendedFlags & skipWorktreeExtendedFlag) == 0 {
-                entries.append(entryStat)
-            }
-
-            offset += 1
-            if version != 4 {
-                let entryLength = offset - entryStart
-                let padding = (8 - (entryLength % 8)) % 8
-                offset += padding
-            }
-        }
-
-        let checksum = bytes[(bytes.count - 20)..<bytes.count].map { String(format: "%02x", $0) }.joined()
-        return GitIndexSnapshot(
-            entries: entries,
-            signature: checksum,
-            contentSignature: gitIndexContentSignature(entries: contentEntries)
-        )
-    }
-
-    private nonisolated static func gitIndexContentSignature(entries: [GitIndexEntryStat]) -> String {
-        var hash: UInt64 = 14_695_981_039_346_656_037
-
-        func appendByte(_ byte: UInt8) {
-            hash ^= UInt64(byte)
-            hash = hash &* 1_099_511_628_211
-        }
-
-        func appendUInt32(_ value: UInt32) {
-            appendByte(UInt8((value >> 24) & 0xff))
-            appendByte(UInt8((value >> 16) & 0xff))
-            appendByte(UInt8((value >> 8) & 0xff))
-            appendByte(UInt8(value & 0xff))
-        }
-
-        func appendString(_ value: String) {
-            for byte in value.utf8 {
-                appendByte(byte)
-            }
-        }
-
-        appendUInt32(UInt32(truncatingIfNeeded: entries.count))
-        for entry in entries {
-            appendString(entry.path)
-            appendByte(0)
-            appendUInt32(entry.mode)
-            appendByte(0)
-            appendString(entry.objectID)
-            appendByte(0)
-        }
-        return String(format: "%016llx", CUnsignedLongLong(hash))
-    }
-
-    private nonisolated static func gitlinkWorktreeCommit(
-        parentRepository: ResolvedGitRepository,
-        gitlinkPath: String
-    ) -> String? {
-        let gitlinkURL = URL(fileURLWithPath: parentRepository.workTreeRoot)
-            .appendingPathComponent(gitlinkPath)
-            .standardizedFileURL
-        guard let submoduleRepository = resolveGitRepository(containing: gitlinkURL.path),
-              submoduleRepository.workTreeRoot == gitlinkURL.path else {
-            return nil
-        }
-        return gitCurrentCommit(repository: submoduleRepository)
-    }
-
-    private nonisolated static func gitCurrentCommit(repository: ResolvedGitRepository) -> String? {
-        let headURL = URL(fileURLWithPath: repository.gitDirectory).appendingPathComponent("HEAD")
-        guard let contents = try? String(contentsOf: headURL, encoding: .utf8) else {
-            return nil
-        }
-        let trimmed = contents.trimmingCharacters(in: .whitespacesAndNewlines)
-        let refPrefix = "ref: "
-        let value: String
-        if trimmed.hasPrefix(refPrefix) {
-            let refName = String(trimmed.dropFirst(refPrefix.count))
-            guard !refName.isEmpty, let refValue = gitRefValue(repository: repository, refName: refName) else {
-                return nil
-            }
-            value = refValue
-        } else {
-            value = trimmed
-        }
-
-        let normalized = value.trimmingCharacters(in: .whitespacesAndNewlines).lowercased()
-        guard normalized.count == 40,
-              normalized.allSatisfy({ $0.isHexDigit }) else {
-            return nil
-        }
-        return normalized
-    }
-
-    private nonisolated static func gitIndexComparableMode(for statMode: mode_t) -> UInt32? {
-        let type = statMode & mode_t(S_IFMT)
-        switch type {
-        case mode_t(S_IFREG):
-            return (statMode & 0o111) == 0 ? 0o100644 : 0o100755
-        case mode_t(S_IFLNK):
-            return 0o120000
-        default:
-            return nil
-        }
-    }
-
-    private nonisolated static func gitIndexUInt32Field<T: BinaryInteger>(_ value: T) -> UInt32 {
-        UInt32(truncatingIfNeeded: UInt64(truncatingIfNeeded: value))
-    }
-
-    private nonisolated static func gitIndexFileSignature(indexURL: URL) -> String? {
-        guard let data = try? Data(contentsOf: indexURL), data.count >= 20 else {
-            return nil
-        }
-        return data.suffix(20).map { String(format: "%02x", $0) }.joined()
-    }
-
-    private nonisolated static func readGitIndexV4PathStripLength(
-        _ bytes: [UInt8],
-        offset: inout Int
-    ) -> Int? {
-        guard offset < bytes.count else { return nil }
-        var byte = bytes[offset]
-        offset += 1
-        var value = Int(byte & 0x7f)
-        while (byte & 0x80) != 0 {
-            guard offset < bytes.count else { return nil }
-            // Git's index v4 path compression uses varint.c's encode/decode pair.
-            // Continuation bytes increment the accumulated value before shifting.
-            value += 1
-            byte = bytes[offset]
-            offset += 1
-            value = (value << 7) + Int(byte & 0x7f)
-        }
-        return value
-    }
-
-    private nonisolated static func readBigEndianUInt16(_ bytes: [UInt8], at offset: Int) -> UInt16 {
-        (UInt16(bytes[offset]) << 8) | UInt16(bytes[offset + 1])
-    }
-
-    private nonisolated static func readBigEndianUInt32(_ bytes: [UInt8], at offset: Int) -> UInt32 {
-        (UInt32(bytes[offset]) << 24) |
-            (UInt32(bytes[offset + 1]) << 16) |
-            (UInt32(bytes[offset + 2]) << 8) |
-            UInt32(bytes[offset + 3])
     }
 
     private nonisolated static func fetchWorkspacePullRequestRepoResults(
@@ -4843,132 +3848,6 @@ class TabManager: ObservableObject {
 
     private nonisolated static func decodeJSON<T: Decodable>(_ type: T.Type, from data: Data) -> T? {
         try? JSONDecoder().decode(T.self, from: data)
-    }
-
-    nonisolated static func githubRepositorySlugs(fromGitRemoteVOutput output: String) -> [String] {
-        var slugByRemoteName: [String: String] = [:]
-
-        for line in output.split(whereSeparator: \.isNewline) {
-            let parts = line.split(whereSeparator: \.isWhitespace)
-            guard parts.count >= 3 else { continue }
-
-            let remoteName = String(parts[0])
-            let remoteURL = String(parts[1])
-            let remoteKind = String(parts[2])
-            guard remoteKind == "(fetch)",
-                  let repoSlug = githubRepositorySlug(fromRemoteURL: remoteURL) else {
-                continue
-            }
-
-            slugByRemoteName[remoteName] = repoSlug
-        }
-
-        let orderedRemoteNames = slugByRemoteName.keys.sorted { lhs, rhs in
-            let lhsPriority = githubRemotePriority(lhs)
-            let rhsPriority = githubRemotePriority(rhs)
-            if lhsPriority != rhsPriority {
-                return lhsPriority < rhsPriority
-            }
-            return lhs < rhs
-        }
-
-        var orderedSlugs: [String] = []
-        var seen: Set<String> = []
-        for remoteName in orderedRemoteNames {
-            guard let repoSlug = slugByRemoteName[remoteName],
-                  seen.insert(repoSlug).inserted else {
-                continue
-            }
-            orderedSlugs.append(repoSlug)
-        }
-        return orderedSlugs
-    }
-
-#if DEBUG
-    nonisolated static func workspaceGitMetadataWatchedPathsForTesting(directory: String) -> [String] {
-        workspaceGitMetadataWatchedPaths(for: directory) ?? []
-    }
-
-    nonisolated static func githubRepositorySlugs(fromGitConfigForTesting config: String) -> [String] {
-        githubRepositorySlugs(fromGitRemoteVOutput: gitRemoteVLines(fromConfig: config).joined())
-    }
-
-    nonisolated static func githubRepositorySlugs(directoryForTesting directory: String) -> [String] {
-        guard let repository = resolveGitRepository(containing: directory),
-              let output = gitRemoteVOutput(repository: repository) else {
-            return []
-        }
-        return githubRepositorySlugs(fromGitRemoteVOutput: output)
-    }
-#endif
-
-    #if compiler(>=6.2)
-    @concurrent
-    #endif
-    private nonisolated static func githubRepositorySlugs(directory: String) async -> [String] {
-        guard let repository = resolveGitRepository(containing: directory),
-              let output = gitRemoteVOutput(repository: repository) else {
-            return []
-        }
-        return githubRepositorySlugs(fromGitRemoteVOutput: output)
-    }
-
-    private nonisolated static func githubRemotePriority(_ remoteName: String) -> Int {
-        switch remoteName.lowercased() {
-        case "upstream":
-            return 0
-        case "origin":
-            return 1
-        default:
-            return 2
-        }
-    }
-
-    private nonisolated static func githubRepositorySlug(fromRemoteURL remoteURL: String) -> String? {
-        let trimmed = remoteURL.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard !trimmed.isEmpty else { return nil }
-
-        let githubPrefixes = [
-            "git@github.com:",
-            "ssh://git@github.com/",
-            "https://github.com/",
-            "http://github.com/",
-            "git://github.com/",
-        ]
-        for prefix in githubPrefixes where trimmed.hasPrefix(prefix) {
-            let path = String(trimmed.dropFirst(prefix.count))
-            return normalizedGitHubRepositorySlug(path)
-        }
-
-        guard let url = URL(string: trimmed),
-              let host = url.host?.lowercased(),
-              host == "github.com" else {
-            return nil
-        }
-
-        return normalizedGitHubRepositorySlug(url.path)
-    }
-
-    private nonisolated static func githubRepositorySlug(fromPullRequestURL url: URL) -> String? {
-        guard let host = url.host?.lowercased(),
-              host == "github.com" else {
-            return nil
-        }
-        return normalizedGitHubRepositorySlug(url.path)
-    }
-
-    private nonisolated static func normalizedGitHubRepositorySlug(_ rawPath: String) -> String? {
-        let trimmedPath = rawPath.trimmingCharacters(in: CharacterSet(charactersIn: "/"))
-        guard !trimmedPath.isEmpty else { return nil }
-        let components = trimmedPath.split(separator: "/").map(String.init)
-        guard components.count >= 2 else { return nil }
-        let owner = components[0]
-        var repo = components[1]
-        if repo.hasSuffix(".git") {
-            repo.removeLast(4)
-        }
-        guard !owner.isEmpty, !repo.isEmpty else { return nil }
-        return "\(owner)/\(repo)"
     }
 
     private nonisolated static func debugLogSnippet(_ value: String?) -> String? {
