+++ cmuxTests/TabManagerUnitTests.swift
@@ -613,198 +613,6 @@ final class TabManagerWorkspaceOwnershipTests: XCTestCase {
 
 @MainActor
 final class TabManagerPullRequestProbeTests: XCTestCase {
-    func testGitHubRepositorySlugsPrioritizeUpstreamThenOriginAndDeduplicate() {
-        let output = """
-        origin https://github.com/austinwang/cmux.git (fetch)
-        origin https://github.com/austinwang/cmux.git (push)
-        upstream git@github.com:manaflow-ai/cmux.git (fetch)
-        upstream git@github.com:manaflow-ai/cmux.git (push)
-        backup ssh://git@github.com/manaflow-ai/cmux.git (fetch)
-        mirror https://gitlab.com/manaflow-ai/cmux.git (fetch)
-        """
-
-        XCTAssertEqual(
-            TabManager.githubRepositorySlugs(fromGitRemoteVOutput: output),
-            ["manaflow-ai/cmux", "austinwang/cmux"]
-        )
-    }
-
-    func testGitHubRepositorySlugsFromGitConfigIgnoreInlineComments() {
-        let config = """
-        [remote "origin"] ; user's main fork
-            url = git@github.com:austinwang/cmux.git # main origin
-            fetch = +refs/heads/*:refs/remotes/origin/*
-        [remote "upstream"] # canonical repo
-            url = https://github.com/manaflow-ai/cmux.git ; upstream source
-            fetch = +refs/heads/*:refs/remotes/upstream/*
-        """
-
-        XCTAssertEqual(
-            TabManager.githubRepositorySlugs(fromGitConfigForTesting: config),
-            ["manaflow-ai/cmux", "austinwang/cmux"]
-        )
-    }
-
-    func testGitHubRepositorySlugsFromGitConfigUnquotesUrlValues() {
-        let config = """
-        [remote "origin"] ; user's main fork
-            url = "git@github.com:austinwang/cmux.git" # main origin
-            fetch = +refs/heads/*:refs/remotes/origin/*
-        [remote "upstream"] # canonical repo
-            url = "https://github.com/manaflow-ai/cmux.git" ; upstream source
-            fetch = +refs/heads/*:refs/remotes/upstream/*
-        """
-
-        XCTAssertEqual(
-            TabManager.githubRepositorySlugs(fromGitConfigForTesting: config),
-            ["manaflow-ai/cmux", "austinwang/cmux"]
-        )
-    }
-
-    func testGitHubRepositorySlugsFromGitConfigUsesLastRemoteURLValue() {
-        let config = """
-        [remote "origin"]
-            url = https://github.com/old-owner/old-repo.git
-            url = https://github.com/manaflow-ai/cmux.git
-        """
-
-        XCTAssertEqual(
-            TabManager.githubRepositorySlugs(fromGitConfigForTesting: config),
-            ["manaflow-ai/cmux"]
-        )
-    }
-
-    func testGitHubRepositorySlugsFromGitConfigReadsIncludedConfigFiles() throws {
-        let repoURL = FileManager.default.temporaryDirectory.appendingPathComponent(
-            "cmux-git-config-includes-\(UUID().uuidString)",
-            isDirectory: true
-        )
-        let gitURL = repoURL.appendingPathComponent(".git", isDirectory: true)
-        try FileManager.default.createDirectory(at: gitURL, withIntermediateDirectories: true)
-        defer {
-            try? FileManager.default.removeItem(at: repoURL)
-        }
-
-        try "ref: refs/heads/main\n".write(
-            to: gitURL.appendingPathComponent("HEAD"),
-            atomically: true,
-            encoding: .utf8
-        )
-        try """
-        [include]
-            path = remotes.inc
-        [includeIf "gitdir:\(gitURL.path)/**"]
-            path = conditional-remotes.inc
-        """.write(
-            to: gitURL.appendingPathComponent("config"),
-            atomically: true,
-            encoding: .utf8
-        )
-        try """
-        [remote "origin"]
-            url = "git@github.com:austinwang/cmux.git" # user's main fork
-        """.write(
-            to: gitURL.appendingPathComponent("remotes.inc"),
-            atomically: true,
-            encoding: .utf8
-        )
-        try """
-        [remote "upstream"]
-            url = https://github.com/manaflow-ai/cmux.git ; canonical repo
-        """.write(
-            to: gitURL.appendingPathComponent("conditional-remotes.inc"),
-            atomically: true,
-            encoding: .utf8
-        )
-
-        XCTAssertEqual(
-            TabManager.githubRepositorySlugs(directoryForTesting: repoURL.path),
-            ["manaflow-ai/cmux", "austinwang/cmux"]
-        )
-    }
-
-    func testGitHubRepositorySlugsFromGitConfigAppliesIncludesInPlace() throws {
-        let repoURL = FileManager.default.temporaryDirectory.appendingPathComponent(
-            "cmux-git-config-include-order-\(UUID().uuidString)",
-            isDirectory: true
-        )
-        let gitURL = repoURL.appendingPathComponent(".git", isDirectory: true)
-        try FileManager.default.createDirectory(at: gitURL, withIntermediateDirectories: true)
-        defer {
-            try? FileManager.default.removeItem(at: repoURL)
-        }
-
-        try "ref: refs/heads/main\n".write(
-            to: gitURL.appendingPathComponent("HEAD"),
-            atomically: true,
-            encoding: .utf8
-        )
-        try """
-        [include]
-            path = remotes.inc
-        [remote "origin"]
-            url = https://github.com/manaflow-ai/cmux.git
-        """.write(
-            to: gitURL.appendingPathComponent("config"),
-            atomically: true,
-            encoding: .utf8
-        )
-        try """
-        [remote "origin"]
-            url = https://github.com/old-owner/old-repo.git
-        """.write(
-            to: gitURL.appendingPathComponent("remotes.inc"),
-            atomically: true,
-            encoding: .utf8
-        )
-
-        XCTAssertEqual(
-            TabManager.githubRepositorySlugs(directoryForTesting: repoURL.path),
-            ["manaflow-ai/cmux"]
-        )
-    }
-
-    func testGitHubRepositorySlugsFromGitConfigTreatsTrailingSlashGitdirAsRecursive() throws {
-        let parentURL = FileManager.default.temporaryDirectory.appendingPathComponent(
-            "cmux-git-config-recursive-include-\(UUID().uuidString)",
-            isDirectory: true
-        )
-        let repoURL = parentURL
-            .appendingPathComponent("teams", isDirectory: true)
-            .appendingPathComponent("cmux", isDirectory: true)
-        let gitURL = repoURL.appendingPathComponent(".git", isDirectory: true)
-        try FileManager.default.createDirectory(at: gitURL, withIntermediateDirectories: true)
-        defer {
-            try? FileManager.default.removeItem(at: parentURL)
-        }
-
-        try "ref: refs/heads/main\n".write(
-            to: gitURL.appendingPathComponent("HEAD"),
-            atomically: true,
-            encoding: .utf8
-        )
-        try """
-        [includeIf "gitdir:\(parentURL.path)/"]
-            path = recursive-remotes.inc
-        """.write(
-            to: gitURL.appendingPathComponent("config"),
-            atomically: true,
-            encoding: .utf8
-        )
-        try """
-        [remote "upstream"]
-            url = https://github.com/manaflow-ai/cmux.git
-        """.write(
-            to: gitURL.appendingPathComponent("recursive-remotes.inc"),
-            atomically: true,
-            encoding: .utf8
-        )
-
-        XCTAssertEqual(
-            TabManager.githubRepositorySlugs(directoryForTesting: repoURL.path),
-            ["manaflow-ai/cmux"]
-        )
-    }
 
     func testPreferredPullRequestPrefersOpenOverMergedAndClosed() {
         let candidates = [
